Extract OpenAI protocol formatting and request parsing from openai_compat
openai_compat.py grew into a 1850-line module mixing SSE protocol
formatting, request/model-identity parsing, and execution orchestration.
Split it into three pieces with openai_compat keeping only route handlers
and orchestration:

- api/openai_streaming_protocol.py: per-stream completion state
  (CompletionStreamState), SSE chunk assembly and framing, tool-call
  trace encoding, error-body shaping, error-string detection, and the
  finalizer-aware response classes. Input is core stream chunk events,
  output is wire format.
- api/openai_request_parsing.py: chat-completion body parsing, OpenAI
  message conversion to MindRoom prompt/thread-history inputs, session-ID
  derivation, and agent/team model-identity validation. Pure given the
  request payload and the config snapshot.

No behavior change for OpenAI-protocol clients: same JSON shapes, SSE
framing, and error responses. tests/test_openai_compat.py passes with a
single import update (_extract_content_text moved); underscore aliases in
openai_compat keep existing patch targets stable. tach.toml gains entries
for the new modules and drops openai_compat deps that are now
TYPE_CHECKING-only or stale.
